requirements.txt
ultralytics==8.0.168             # YOLOv8 stable
deep-sort-realtime>=1.3         # Tracker interface stable from 1.3
deap>=1.3
numpy
pandas
opencv-python
matplotlib
tqdm
PyYAML
